Crate hacspec_linalg

Functions

identity

Generates an identity matrix. If the matrix is not square, the largest square submatrix (starting at the first row and column) is set to the identity while all other entries are set to zero.

index

Gets the index of a matrix

new

Generate new matrix using rows, cols and a seq.

ones

Generate a n*m matrix filled with ones.

repeat

Generate a n*m matrix filled with a given scalar

scale

Scale a matrix with a given scalar

slice

Returns a matrix slice, given a Matrix and two Dims (pairs of usize), first Dims pair representing the starting point and second Dims pair the dimensions

transpose

Transposes a matrix

zeros

Generate a n*m matrix filled with zeros.

Type Definitions

DimType

Dims

MatRes

Matrix

ScalRes

Scalar